notification/requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.38.0
redis==7.0.1
aio-pika==9.5.7
python-dotenv
tenacity==9.1.2
httpx==0.28.1
email-validator==2.3.0
structlog==25.5.0
aiosmtplib==5.0.0
pyfcm==2.1.0
webpush==1.0.6
pydantic-settings==2.12.0
pytest==9.0.0